pyproject.toml
[project]
name = "rename-image-files"
version = "0.2.0-a1"
description = "Rename image files using vision language models"
authors = [{ name = "Oliver Steele", email = "steele@osteele.com" }]
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }

dependencies = ["click", "exifread", "llm", "llm-gemini", "rich", "pillow", "pillow-heif"]

[dependency-groups]
dev = ["pytest", "pytest-cov", "ruff", "pyright"]

[project.scripts]
rename-image-files = "rename_image_files.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
  "E",  # pycodestyle errors
  "F",  # Pyflakes rules
  "I",  # isort import sorting
  "UP", # upgrade syntax to newer Python versions
]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
reportMissingTypeStubs = false
exclude = [".venv"]
reportUnusedImport = false

[[tool.pyright.overrides]]
module = ["exifread", "llm", "llm.*", "rich", "rich.*"]
ignore_missing_imports = true
